konfigsets: Read files from specific namespace
Any konfigset file that are declared, is checked against only the
default namespace.  However, a configmap might be living in a
different namespace.

Therefore, let's add the ability to read the configmap from a
different namespace, which is declared on each konfigset file entry
definitions. In case a namespace is not given, konfig falls back to
the namespace where the konfigset is created.

Additionally, to make it clear in the error logs message, we also
append namespace to the configmap name in the following format:
"<namespace>/<name>"; thus making it clear which namespace and name it
tried to fetch the configmap from in case of errors.

While we are it, fix comment where the pattern k8s://static is
specified which is nonsensical, given that static content are defined
as static://.

Signed-off-by: Rafael Silva <rafael@raffs.dev>