konfig: Add command-line args with clap
Clap seems to be a popular command-line interface args library. It
provides macros for annotating structures and its arguments and
automatically renders the arguments, as well and sub-command.

Therefore, let's add in both `konfigd` and `konfigm` binary in order
to create the infra for future commands.  While we are it, let's add a
`--knodename` arguments to `konfigd`, in order to override the name it
choose when register against k8s control plane. As default, it
fallsback to grab the system hostname, maintain backward compability.

Signed-off-by: Rafael Silva <rafael@raffs.dev>